backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.5
pydantic-settings==2.7.1

# PDF Processing
PyMuPDF==1.25.3
pdfplumber==0.11.4
camelot-py[cv]==0.11.0
tabula-py==2.10.0

# OCR
pytesseract==0.3.13
Pillow==11.1.0

# NLP / AI
anthropic==0.44.0
spacy==3.8.4

# Data Processing
numpy==2.2.3
pandas==2.2.3

# Database
supabase==2.13.0
asyncpg==0.30.0
sqlalchemy==2.0.37

# Cache
redis==5.2.1

# Cloud Storage
google-cloud-storage==2.19.0
google-cloud-vision==3.9.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
tenacity==9.0.0
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.3
httpx==0.28.1
